+++ src/lib.rs
@@ -1,7 +1,17 @@
-pub mod runtime;
-pub mod exceptions;
-pub mod parsing;
-pub mod byte_readable;
-pub mod code;
-pub mod instruction;
-pub mod instruction_table;
+pub use runtime::{Machine, InstructionPointer, CallFrame};
+pub use runtime::exceptions as runtime_exceptions;
+pub use parsing::{CodeParser, ConstantParser, ConstantParserTable, RawBytes, RawBytesPointer};
+pub use parsing::exceptions as parsing_exceptions;
+pub use byte_readable::ByteReadable;
+pub use code::{Chunk, Code};
+pub use instruction::{Instruction, InstructionFn, RawInstructionFn};
+pub use instruction_table::InstructionTable;
+pub use exception::{Exception, ExceptionType};
+
+mod runtime;
+mod parsing;
+mod byte_readable;
+mod code;
+mod instruction;
+mod instruction_table;
+mod exception;
