+++ src/parsing/mod.rs
@@ -1,4 +1,8 @@
-pub mod code_parser;
-pub mod constant_parser;
-pub mod parsing_exceptions;
-pub mod raw_bytes;
+pub use code_parser::CodeParser;
+pub use constant_parser::{ConstantParser, ConstantParserTable};
+pub use raw_bytes::{RawBytes, RawBytesPointer};
+
+pub mod exceptions;
+mod raw_bytes;
+mod code_parser;
+mod constant_parser;
