+++ src/runtime/mod.rs
@@ -1,4 +1,10 @@
+pub use call_frame::CallFrame;
+pub use instruction_pointer::InstructionPointer;
+pub use machine::Machine;
+
+
 mod call_frame;
-pub mod instruction_pointer;
-pub mod machine;
+mod instruction_pointer;
+mod machine;
 mod stack;
+pub mod exceptions;
